+++ src/test/java/slt/rest/MealplanControllerTest.java
@@ -25,14 +25,14 @@ class MealplanControllerTest {
         repository = mock(MealplanRepository.class);
         controller = new MealplanController(repository);
         final var userInfo = new UserInfo();
-        userInfo.setUserId(1);
+        userInfo.setUserId(1L);
         ThreadLocalHolder.getThreadLocal().set(userInfo);
     }
 
     @Test
     void getAllMealplans() {
         final var mealplans = List.of(Mealplan.builder().id(1L).title("title").mealtimes(new ArrayList<>()).build());
-        when(repository.getAllMealplans(any(Integer.class))).thenReturn(mealplans);
+        when(repository.getAllMealplans(any(Long.class))).thenReturn(mealplans);
         final var result = controller.getAllMealplans();
         Assertions.assertNotNull(result.getBody());
         Assertions.assertEquals(1, result.getBody().size());
@@ -66,7 +66,7 @@ class MealplanControllerTest {
     @Test
     void deleteMealplan() {
         final var result = controller.deleteMealplan(1L);
-        verify(repository).deleteMealplan(1, 1L);
+        verify(repository).deleteMealplan(1L, 1L);
         Assertions.assertNull(result.getBody());
     }
 }
